Fix compilation error caused by adding content length log in CBLRemoteRequest
Casted to unsigned long to fix the compilation error.

+++ Source/CBLRemoteRequest.m
@@ -84,7 +84,7 @@ typedef enum {
             }
             _request.HTTPBody = body;
             LogTo(RemoteRequest, @"%@:Added HTTP body of size %lu bytes",
-                  self, _request.HTTPBody.length);
+                  self, (unsigned long)_request.HTTPBody.length);
         }
     }
     return self;
